Backend configured, debugged and complete onto frontend

+++ Beginner/To-do list/Server/Modules/Todos/todos.controller.ts
@@ -53,6 +53,16 @@ export const TodoController = (
 
       switch (pathNames[2]) {
         case "create":
+          if (request.method != "POST") {
+            response.writeHead(405);
+            response.end(
+              JSON.stringify({
+                error: "Use POST instead",
+              })
+            );
+            return;
+          }
+
           const createOperation: Todo = await todoService.createTodo(
             userId,
             parsedRequestBody
@@ -62,12 +72,31 @@ export const TodoController = (
           response.end(JSON.stringify(createOperation));
           break;
         case "edit":
+          if (request.method != "PATCH") {
+            response.writeHead(405);
+            response.end(
+              JSON.stringify({
+                error: "Use PATCH instead",
+              })
+            );
+            return;
+          }
+
           const editOperation = await todoService.editTodo(parsedRequestBody);
 
           response.writeHead(201);
           response.end(JSON.stringify(editOperation));
           break;
         case "get":
+          if (request.method != "GET") {
+            response.writeHead(405);
+            response.end(
+              JSON.stringify({
+                error: "Use GET instead",
+              })
+            );
+            return;
+          }
           const searchParams = requestUrl.searchParams,
             type = searchParams.get("type");
 
@@ -75,7 +104,7 @@ export const TodoController = (
             response.writeHead(404);
             response.end(
               JSON.stringify({
-                error: "Provide type and user id in the search params",
+                error: "Provide type in the search params",
               })
             );
             return;
@@ -106,11 +135,31 @@ export const TodoController = (
           }
           break;
         case "delete":
+          if (request.method != "DELETE") {
+            response.writeHead(405);
+            response.end(
+              JSON.stringify({
+                error: "Use DELETE instead",
+              })
+            );
+            return;
+          }
+
           const searchDeletionParams = requestUrl.searchParams,
             typeDeletion = searchDeletionParams.get("type");
 
+          if (!typeDeletion) {
+            response.writeHead(400);
+            response.end(
+              JSON.stringify({
+                error: "Provide type in search params",
+              })
+            );
+            return;
+          }
+
           if (typeDeletion == "one") {
-            const todoId = searchDeletionParams.get("todoId");
+            const todoId = searchDeletionParams.get("todoid");
 
             if (!todoId) {
               response.writeHead(404);
@@ -126,11 +175,18 @@ export const TodoController = (
 
             response.writeHead(204);
             response.end();
-          } else {
+          } else if (typeDeletion == "all") {
             await todoService.deleteTodos(userId);
 
             response.writeHead(204);
             response.end();
+          } else {
+            response.writeHead(400);
+            response.end(
+              JSON.stringify({
+                error: "Invalid type value specify all or one",
+              })
+            );
           }
           break;
       }
